Optimize sending of album artwork
We now create an SHA256 checksum of each file, set this hash as the
picture value in the object-to-be-returned, and then return only one
copy of a picture and its associated checksum. The /api/info now returns
data similar to:

    {
      art: {"4f0f80979da058b84d6672464dffa862567e082a87af1fffdee101184":
      ...},
      tracks: [
        {..., picture: "4f0f80979da058b84d6672464dffa862567e082a87af1fffdee101184" }
        ...
      ]
    }

Clients may also provide an array of SHA256 strings as the "art"
querystring. Doing so will prevent the server from sending that specific
SHA256 artwork to the client.